backend/requirements.txt
# ---- FastAPI
fastapi==0.111.0
uvicorn==0.30.1
httptools==0.6.4

# ---- DB & Vector
psycopg2-binary==2.9.9
pgvector==0.2.5

# ---- LangChain stack
langchain-community
langchain-ollama>=0.1.0
langgraph
flashrank
sentence-transformers
requests

# ---- Loaders & utils
langchain-docling
pypdf
python-dotenv
python-multipart

# ---- Whisper
faster-whisper
soundfile

# ---- OCR
pytesseract
Pillow

# ---- RAGAS & friends
ragas==0.1.17
datasets>=2.19.0
pandas>=2.1
tqdm

langchain-openai>=0.2.0